Clean up SQL, refactor serverName config because it was mispelled.
Now using all prepared statements the correct way. Changed the connection to a method instead of a variable so it refreshes every time it attempts to connect to the database. Changed all the connections and statements to be inside of a try catch statement so they automatically close (to avoid leaks).

Took 21 minutes